requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
easyocr==1.7.2
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
gunicorn==25.3.0
h11==0.16.0
h5py==3.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.2
idna==3.11
ImageIO==2.37.3
imutils==0.5.4
Jinja2==3.1.6
lazy-loader==0.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pyclipper==1.4.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
PyYAML==6.0.3
rich==14.3.3
scikit-image==0.26.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
super-image==0.2.0
sympy==1.14.0
tifffile==2026.3.3
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
